+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Airline - Change Log
 
+## 2.8.4
+
+- Restriction Improvements
+    - New `@NoOptionLikeValues` restriction for explicitly rejecting values that look like options i.e. start with
+      `-` or `--` but didn't match a defined option e.g. user typos (#112)
+
 ## 2.8.3
 
 - Parser Improvements